move metrics code from response to metrics model
