+++ db/migrate/20190425194052_create_metrics.rb
@@ -1,0 +1,8 @@
+class CreateMetrics < ActiveRecord::Migration
+  def change
+    create_table :metrics do |t|
+
+      t.timestamps null: false
+    end
+  end
+end
